+++ lib/Crypto/Cipher/_mode_kwp.py
@@ -1,50 +1,13 @@
 import struct
-from collections import deque
 
 from types import ModuleType
 from typing import Union
 
-from Crypto.Util.strxor import strxor
+from ._mode_kw import W, W_inverse
 
 
-def W(cipher: ModuleType,
-      plaintext: Union[bytes, bytearray]) -> bytes:
-
-    S = [plaintext[i:i+8] for i in range(0, len(plaintext), 8)]
-    n = len(S)
-    s = 6 * (n - 1)
-    A = S[0]
-    R = deque(S[1:])
-
-    for t in range(1, s + 1):
-        t_64 = struct.pack('>Q', t)
-        ct = cipher.encrypt(A + R.popleft())
-        A = strxor(ct[:8], t_64)
-        R.append(ct[8:])
-
-    return A + b''.join(R)
-
-
-def W_inverse(cipher: ModuleType,
-              ciphertext: Union[bytes, bytearray]) -> bytes:
-
-    C = [ciphertext[i:i+8] for i in range(0, len(ciphertext), 8)]
-    n = len(C)
-    s = 6 * (n - 1)
-    A = C[0]
-    R = deque(C[1:])
-
-    for t in range(s, 0, -1):
-        t_64 = struct.pack('>Q', t)
-        pt = cipher.decrypt(strxor(A, t_64) + R.pop())
-        A = pt[:8]
-        R.appendleft(pt[8:])
-
-    return A + b''.join(R)
-
-
-class KWMode(object):
-    """Key Wrap (KW) mode.
+class KWPMode(object):
+    """Key Wrap with Padding (KWP) mode.
 
     This is a deterministic Authenticated Encryption (AE) mode
     for protecting cryptographic keys. See `NIST SP800-38F`_.
@@ -66,7 +29,7 @@ class KWMode(object):
 
         self.block_size = factory.block_size
         if self.block_size != 16:
-            raise ValueError("Key Wrap mode is only available for ciphers"
+            raise ValueError("Key Wrap with Padding mode is only available for ciphers"
                              " that operate on 128 bits blocks")
 
         self._factory = factory
@@ -79,8 +42,6 @@ class KWMode(object):
         Args:
           plaintext:
             The cryptographic key to wrap.
-            It must be at least 16 bytes long, and its length
-            must be a multiple of 8.
 
         Returns:
             The wrapped key.
@@ -89,17 +50,22 @@ class KWMode(object):
         if self._done:
             raise ValueError("The cipher cannot be used more than once")
 
-        if len(plaintext) % 8:
-            raise ValueError("The plaintext must have length multiple of 8 bytes")
+        if len(plaintext) == 0:
+            raise ValueError("The plaintext must be at least 1 byte")
 
-        if len(plaintext) < 16:
-            raise ValueError("The plaintext must be at least 16 bytes long")
-
-        if len(plaintext) >= 2**32:
+        if len(plaintext) >= 2 ** 32:
             raise ValueError("The plaintext is too long")
 
-        res = W(self._cipher, b'\xA6\xA6\xA6\xA6\xA6\xA6\xA6\xA6' + plaintext)
-        self._done = True
+        padlen = (8 - len(plaintext)) % 8
+        padded = plaintext + b'\x00' * padlen
+
+        AIV = b'\xA6\x59\x59\xA6' + struct.pack('>I', len(plaintext))
+
+        if len(padded) == 8:
+            res = self._cipher.encrypt(AIV + padded)
+        else:
+            res = W(self._cipher, AIV + padded)
+
         return res
 
     def unseal(self, ciphertext: Union[bytes, bytearray]) -> bytes:
@@ -108,7 +74,7 @@ class KWMode(object):
         Args:
           ciphertext:
             The cryptographic key to unwrap.
-            It must be at least 24 bytes long, and its length
+            It must be at least 16 bytes long, and its length
             must be a multiple of 8.
 
         Returns:
@@ -124,21 +90,32 @@ class KWMode(object):
         if len(ciphertext) % 8:
             raise ValueError("The ciphertext must have length multiple of 8 bytes")
 
-        if len(ciphertext) < 24:
+        if len(ciphertext) < 16:
             raise ValueError("The ciphertext must be at least 24 bytes long")
 
-        pt = W_inverse(self._cipher, ciphertext)
+        if len(ciphertext) == 16:
+            S = self._cipher.decrypt(ciphertext)
+        else:
+            S = W_inverse(self._cipher, ciphertext)
 
-        if pt[:8] != b'\xA6\xA6\xA6\xA6\xA6\xA6\xA6\xA6':
-            raise ValueError("Incorrect integrity check value")
-        self._done = True
+        if S[:4] != b'\xA6\x59\x59\xA6':
+            raise ValueError("Incorrect decryption")
 
-        return pt[8:]
+        Plen = struct.unpack('>I', S[4:8])[0]
+
+        padlen = len(S) - 8 - Plen
+        if padlen < 0 or padlen > 7:
+            raise ValueError("Incorrect decryption")
+
+        if S[len(S) - padlen:] != b'\x00' * padlen:
+            raise ValueError("Incorrect decryption")
+
+        return S[8:len(S) - padlen]
 
 
-def _create_kw_cipher(factory: ModuleType,
-                      **kwargs: Union[bytes, bytearray]) -> KWMode:
-    """Create a new block cipher in Key Wrap mode.
+def _create_kwp_cipher(factory: ModuleType,
+                       **kwargs: Union[bytes, bytearray]) -> KWPMode:
+    """Create a new block cipher in Key Wrap with Padding mode.
 
     Args:
       factory:
@@ -155,4 +132,4 @@ def _create_kw_cipher(factory: ModuleType,
     except KeyError as e:
         raise TypeError("Missing parameter:" + str(e))
 
-    return KWMode(factory, key)
+    return KWPMode(factory, key)